+++ src/Symfony/Component/Workflow/EventListener/ExpressionLanguage.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Workflow\EventListener;
 
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Authorization\ExpressionLanguage as BaseExpressionLanguage;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Workflow\Exception\RuntimeException;
@@ -22,15 +23,32 @@ use Symfony\Component\Workflow\Exception\RuntimeException;
  */
 class ExpressionLanguage extends BaseExpressionLanguage
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function __construct(CacheItemPoolInterface $cache = null, array $providers = [])
+    {
+        parent::__construct($cache, $providers);
+
+        foreach (['is_anonymous', 'is_authenticated', 'is_fully_authenticated', 'is_granted', 'is_remember_me'] as $name) {
+            if (!isset($this->functions[$name])) {
+                continue;
+            }
+            $compiler = $this->functions[$name]['compiler'];
+            $evaluator = $this->functions[$name]['evaluator'];
+            $this->register($name, $compiler, function (array $variables, $attributes, $object = null) use ($name, $evaluator) {
+                if (!isset($variables['auth_checker']) || !$variables['auth_checker'] instanceof AuthorizationCheckerInterface) {
+                    throw new RuntimeException(sprintf('"%s" cannot be used as the SecurityBundle is not registered in your application.', $name));
+                }
+
+                return $evaluator($variables, $attributes, $object);
+            });
+        }
+    }
+
     protected function registerFunctions()
     {
         parent::registerFunctions();
-
-        $this->register('is_granted', function ($attributes, $object = 'null') {
-            return sprintf('$auth_checker->isGranted(%s, %s)', $attributes, $object);
-        }, function (array $variables, $attributes, $object = null) {
-            return $variables['auth_checker']->isGranted($attributes, $object);
-        });
 
         $this->register('is_valid', function ($object = 'null', $groups = 'null') {
             return sprintf('0 === count($validator->validate(%s, null, %s))', $object, $groups);