+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Added context to `TransitionException` and its child classes whenever they are thrown in `Workflow::apply()`
  * Added `Registry::has()` to check if a workflow exists
  * Added support for `$context[Workflow::DISABLE_ANNOUNCE_EVENT] = true` when calling `workflow->apply()` to not fire the announce event
+ * Made `$tokenStorage`,  `$authorizationChecker`, `$trustResolver` optional constructor parameters of `GuardListener`
 
 5.0.0
 -----